Dev-be 변경사항 최신화

See merge request s09-webmobile1-sub2/S09P12D205!23

+++ backend/src/main/java/com/ssafy/ssap/dto/QuestionCreateDto.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 @Getter
 public class QuestionCreateDto {
 
-    private long userNo;
+    private int userNo;
     private String title;
     private String content;
     private String category;
